Reworking internal RPC
1. Moved storage of XMLRPC methods from  XmlRpcServer into XmlRpcMethods structure.
2. Refactored most methods to use XmlRpcMethods structure instead of
   full fledged server.
3. Refactored internal XMLRPC calls on rosmaster. Now they do go
   directly to RPCManager without involvement of any network code.
4. Added debug timings for HttpServerConnection.
5. master now uses a single RPCManager for both serving master requests
   and rosout requests.

+++ src/roscore/master_handler.cpp
@@ -185,7 +185,7 @@ ReturnStruct MasterHandler::registerSubscriber(const RequesterInfo& requesterInf
   rtn.statusCode = 1;
 
   rtn.value = RpcValue::Array(publishers.size());
-  for (int i = 0; i < publishers.size(); i++) {
+  for (size_t i = 0; i < publishers.size(); i++) {
     if (publishers[i]) {
       std::string strUrl;
       if (publishers[i] != ref) {
@@ -243,7 +243,7 @@ ReturnStruct MasterHandler::registerPublisher(const RequesterInfo& requesterInfo
   rtn.statusMessage = ss.str();
   rtn.statusCode = 1;
   rtn.value = RpcValue::Array(subscribers.size());
-  for (int i = 0; i < subscribers.size(); i++) {
+  for (size_t i = 0; i < subscribers.size(); i++) {
     if (!subscribers[i])
       continue;
     network::URL url = m_resolver.resolveAddressFor(subscribers[i], requesterInfo.clientAddress, requesterInfo.localAddress);
